More updates to crafting system
Added new Item DFN tags:
	usesleft=# 	// The current amount of uses remaining for an item with limited use
	maxuses=# 	// The maximum amount of uses for an item with limted use
Added new JS Item properties to match DFN tags:
	.usesLeft
	.maxUses
Updated scripts to make use of .usesLeft property instead of custom "UsesLeft" tag
Updated DFNs to include usesleft and maxuses tags and uses_left_tooltip script for all tools with limited use
Crafting and gathering tools bought from NPC vendors will now support a random and limited amount of uses (between what is defined as "current" and "max" uses)
Crafting and gathering tools created by player crafters will support a limited amount of uses based on the quality of the item crafted

+++ source/quantityfuncs.cpp
@@ -56,7 +56,7 @@ UI32 GetSubItemAmount( CItem *p, UI16 realID, UI16 realColour, bool colorCheck =
 	{
 		if( ValidateObject( i ) )
 		{
-			if( i->GetType() == IT_CONTAINER || i->GetType() == IT_LOCKEDCONTAINER )
+			if( i->GetID() != realID && ( i->GetType() == IT_CONTAINER || i->GetType() == IT_LOCKEDCONTAINER ))
 				total += GetSubItemAmount( i, realID, realColour );
 			else if( i->GetID() == realID && ( !colorCheck || ( colorCheck && i->GetColour() == realColour )))
 				total += i->GetAmount();
@@ -94,7 +94,7 @@ UI32 DeleteSubItemAmount( CItem *p, UI32 amount, UI16 realID, UI16 realColour )
 	{
 		if( ValidateObject( i ) )
 		{
-			if( i->GetType() == IT_CONTAINER || i->GetType() == IT_LOCKEDCONTAINER ) // Is item an pack or container?
+			if( i->GetID() != realID && ( i->GetType() == IT_CONTAINER || i->GetType() == IT_LOCKEDCONTAINER )) // Is item an pack or container?
 				amtDeleted += DeleteSubItemAmount( i, total, realID, realColour );
 			else if( i->GetID() == realID && i->GetColour() == realColour )
 			{
